security and preauthorize

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/com/pragma/bootcamp/r2dbc/RoleRepository.java
@@ -4,6 +4,6 @@ import com.pragma.bootcamp.r2dbc.entity.RoleEntity;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Mono;
 
-public interface RoleRepository extends ReactiveCrudRepository<RoleEntity, Integer> {
+public interface RoleRepository extends ReactiveCrudRepository<RoleEntity, Long> {
     Mono<RoleEntity> findByName(String name);
 }
