+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/com/pragma/bootcamp/r2dbc/entity/RoleEntity.java
@@ -1,10 +1,6 @@
 package com.pragma.bootcamp.r2dbc.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
 
@@ -17,6 +13,6 @@ import org.springframework.data.relational.core.mapping.Table;
 public class RoleEntity {
 
     @Id
-    private Integer id;
+    private Long id;
     private String name;
 }
